functionalities remove and update added but not working bc duplicates

+++ src/main/java/tsi/ensg/jee/hibernate/colloque/metier/Participant.java
@@ -22,7 +22,7 @@ public class Participant {
     //@Column(name = "id_parti")
 
 
-    public long id_parti;
+    public int id_parti;
     @Column(nullable = false)
     private int num_pers;
     @Column(nullable = false)
@@ -114,7 +114,7 @@ public class Participant {
     public long getId() {
         return id_parti;
     }
-    public void setId(long id_parti){
+    public void setId(int id_parti){
         this.id_parti = id_parti;
     }
