+++ src/main/java/tsi/ensg/jee/hibernate/colloque/services/EvenementService.java
@@ -32,7 +32,7 @@ public class EvenementService {
         session.close();
     }
 
-    public boolean deleteEvenement(Long id) {
+    public boolean deleteEvenement(int id) {
         Evenement evenementToDelete = this.get(id);
         Session session = sessionFactory.openSession();
         session.beginTransaction();
@@ -42,7 +42,7 @@ public class EvenementService {
         return true;
     }
 
-    public Evenement get(long id) {
+    public Evenement get(int id) {
         Session session = sessionFactory.openSession();
         session.beginTransaction();
         Evenement result = session.get(Evenement.class, id);
